Last minute adjustments

+++ space_bases/RocketInputDialog.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
+using System.Diagnostics;
 using System.Drawing;
 using System.Linq;
 using System.Text;
@@ -48,6 +49,9 @@ namespace space_bases
             var Engine1 = Engine1Field.Text;
             var Engine2 = Engine2Field.Text;
             var Engine3 = Engine3Field.Text;
+            var Engine1Count = (int)Engine1Counter.Value;
+            var Engine2Count = (int)Engine2Counter.Value;
+            var Engine3Count = (int)Engine3Counter.Value;
 
             if (RocketName != "" && TrajectoryType != "" && Engine1 != "")
             {
@@ -67,30 +71,60 @@ namespace space_bases
                 db.Rockets.Add(rocket);
                 this.mainWindow.Agency.SpaceAgency = true;
 
+                if (this.db.Engines.Where(e => e.EngineName == Engine1).Count() == 0)
+                {
+                    MessageBox.Show("Engine 1 does not exist", "Error", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    this.db.Rockets.Remove(rocket);
+                    return;
+                }
                 var power1 = new Power
                 {
                     RocketName = RocketName,
-                    EngineName = Engine1
+                    EngineName = Engine1,
+                    Count = Engine1Count
                 };
                 db.Powers.Add(power1);
 
-                var power2 = new Power
-                {
-                    RocketName = RocketName,
-                    EngineName = Engine2
-                };
+
+                Power power2 = null;
                 if (Engine2 != "")
                 {
+                    if(this.db.Engines.Where(e => e.EngineName == Engine2).Count() == 0)
+                    {
+                        MessageBox.Show("Engine 2 does not exist", "Error", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                        this.db.Rockets.Remove(rocket);
+                        db.Powers.Remove(power1);
+                        return;
+                    }
+                    power2 = new Power
+                    {
+                        RocketName = RocketName,
+                        EngineName = Engine2,
+                        Count = Engine2Count
+                    };
                     db.Powers.Add(power2);
                 }
 
-                var power3 = new Power
-                {
-                    RocketName = RocketName,
-                    EngineName = Engine2
-                };
+                Power power3 = null;
                 if (Engine3 != "")
                 {
+                    if (this.db.Engines.Where(e => e.EngineName == Engine3).Count() == 0)
+                    {
+                        MessageBox.Show("Engine 3 does not exist", "Error", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                        this.db.Rockets.Remove(rocket);
+                        db.Powers.Remove(power1);
+                        if (Engine2 != "" && power2 != null)
+                        {
+                            db.Powers.Remove(power2);
+                        }
+                        return;
+                    }
+                    power3 = new Power
+                    {
+                        RocketName = RocketName,
+                        EngineName = Engine3,
+                        Count = Engine3Count
+                    };
                     db.Powers.Add(power3);
                 }
 
@@ -99,17 +133,18 @@ namespace space_bases
                     this.db.SaveChanges();
                     this.mainWindow.IsSpaceAgency = true;
                     this.Close();
+                    this.mainWindow.loadHomePage();
                 }
                 catch (Exception ex)
                 {
                     MessageBox.Show("Error: " + ex.Message, "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
                     this.db.Rockets.Remove(rocket);
                     this.db.Powers.Remove(power1);
-                    if (Engine2 != "")
+                    if (Engine2 != "" && power2 != null)
                     {
                         this.db.Powers.Remove(power2);
                     }
-                    if (Engine3 != "")
+                    if (Engine3 != "" && power3 != null)
                     {
                         this.db.Powers.Remove(power3);
                     }
